pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "theaios-agent-monitor"
version = "0.1.0"
description = "Governance-first observability for AI agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Charafeddine Mouzouni", email = "charafeddine@cohorte.co" },
]
keywords = ["agent-monitoring", "observability", "ai-governance", "kill-switch", "compliance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
guardrails = ["theaios-guardrails>=0.1"]
otel = ["opentelemetry-api>=1.20", "opentelemetry-sdk>=1.20"]
webhook = ["httpx>=0.27"]
all = ["theaios-agent-monitor[guardrails,otel,webhook]"]
dev = ["pytest>=8.0", "ruff>=0.4", "mypy>=1.10", "types-pyyaml>=6.0"]

[project.scripts]
agent-monitor = "theaios.agent_monitor.cli:main"

[project.urls]
Homepage = "https://github.com/Cohorte-ai/agent-monitor"
Documentation = "https://cohorte-ai.github.io/agent-monitor/"
Repository = "https://github.com/Cohorte-ai/agent-monitor"

[tool.hatch.build.targets.wheel]
packages = ["src/theaios"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "opentelemetry.*"
ignore_missing_imports = true
